pyproject.toml
[project]
name = "quake-textures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "scipy>=1.15.2",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
    "gimpformats>=2021.1.3",
    "imageio>=2.37.0",
    "binaryiotools>=2021.0.3",
    "opencv-python>=4.11.0.86",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
